requirements.txt
pandas
snowflake-connector-python
python-dotenv
faker
dbt-snowflake